+++ internal/elasticsearch/watcher/watch/testdata/TestResourceWatch/no_transform/main.tf
@@ -1,0 +1,20 @@
+variable "watch_id" {
+  type = string
+}
+
+provider "elasticstack" {
+  elasticsearch {}
+}
+
+resource "elasticstack_elasticsearch_watch" "test" {
+  watch_id = var.watch_id
+  active   = true
+
+  trigger   = jsonencode({ schedule = { cron = "0 0/2 * * * ?" } })
+  input     = jsonencode({ simple = { name = "example" } })
+  condition = jsonencode({ never = {} })
+  actions   = jsonencode({ log = { logging = { level = "info", text = "example logging text" } } })
+  metadata  = jsonencode({ example_key = "example_value" })
+
+  throttle_period_in_millis = 10000
+}